% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_p_value.R
\name{boot_p_value}
\alias{boot_p_value}
\title{Computing p-values from Bootstrap Confidence Intervals}
\usage{
boot_p_value(
  boot_object,
  ci_method = "bca",
  index = 1,
  null_value = 0,
  alternative = c("two-sided", "greater", "less")[1],
  var_adjust = 1,
  alpha_max = 1,
  alpha_min = 10^-5,
  tolerance = 1e-04,
  max_evaluations = 30,
  verbose = FALSE
)
}
\arguments{
\item{boot_object}{The result of a call to \link[boot:boot]{boot::boot}}

\item{ci_method}{A \code{character} scalar containing the method used to
compute bootstrap confidence intervals}

\item{index}{A \code{numeric} scalar indicating which element of the result
to use in computing a p-value.}

\item{null_value}{A \code{numeric} scalar indicating the null value of the
estimand of interest}

\item{alternative}{A \code{character} scalar indicating the type of
alternative hypothesis being tested: either \code{"two-sided"},
\code{"greater"}, or \code{"less"}.}

\item{var_adjust}{A \code{numeric} scalar for adjusting the confidence
interval width: defaults to 1.}

\item{alpha_max}{A \code{numeric} scalar}

\item{alpha_min}{A \code{numeric} scalar}

\item{tolerance}{A \code{numeric} scalar indicating the tolerance for
assessing convergence}

\item{max_evaluations}{A \code{numeric} scalar indicating the maximum number
of iterations to perform}

\item{verbose}{A \code{logical} scalar, indicating whether to only return
the \emph{p}-value (\code{FALSE}: default), or return the results of each
iteration (\code{TRUE})}
}
\value{
If \code{verbose == FALSE} (default), the \emph{p}-value is returned
as a numeric scalar. Otherwise, a \code{list} is returned.
}
\description{
This function computes \emph{p}-values by finding the smallest
\eqn{100(1 - \alpha)\%} bootstrap confidence interval that does not contain
the null value of the estimand.
}
\details{
The \emph{p}-value is the smallest value of \eqn{\alpha} at which \eqn{H_{0}}
is rejected (i.e. the confidence interval does not contain the null value of
the estimand). This function takes a object of class \code{"boot"} and
iteratively finds the \emph{p}-value through a search algorithm.

A \code{warning} is provided when the approximation may be incorrect, or if
the number of bootstrap replicates may not provide sufficient precision.
}
\examples{
\dontrun{
set.seed(12345)
x <- rnorm(n = 50)

boot_mean <-
  boot::boot(
    data = x,
    statistic = function(data, indices){mean(data[indices])},
    R = 10000
  )

t.test(x = x, alternative = "two.sided")$p.value
boot_p_value(boot_object = boot_mean, alternative = "two-sided")

t.test(x = x, alternative = "less")$p.value
boot_p_value(boot_object = boot_mean, alternative = "less")

t.test(x = x, alternative = "greater")$p.value
boot_p_value(boot_object = boot_mean, alternative = "greater")

t.test(x = x, mu = 0.5, alternative = "less")$p.value
boot_p_value(boot_object = boot_mean, null_value = 0.5, alternative = "less")

t.test(x = x, mu = 0.5, alternative = "greater")$p.value
boot_p_value(boot_object = boot_mean, null_value = 0.5, alternative = "greater")

# Warnings should trigger when approximation may be inaccurate.
t.test(x = x, mu = -1, alternative = "less")$p.value
boot_p_value(boot_object = boot_mean, null_value = -1, alternative = "less")

t.test(x = x, mu = -1, alternative = "greater")$p.value
boot_p_value(boot_object = boot_mean, null_value = -1, alternative = "greater")
}

}
