% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_text_proportion.R
\name{in_text_proportion}
\alias{in_text_proportion}
\title{Format proportions and percentages for presentation}
\usage{
in_text_proportion(
  x,
  n,
  output = c("proportion", "percentage")[1],
  output_format = c("p", "xn", "xp", "xnp")[4],
  digits_proportion = 3,
  digits_percent = 1,
  ci_method = NULL,
  level = 0.95,
  ...
)
}
\arguments{
\item{x}{a \code{numeric} \code{vector} for the numerator.}

\item{n}{a \code{numeric} \code{vector} for the denominator. If
\code{length(x) > 1} and \code{length(n) == 1}, the denominator is assumed
to be the same for all values of \code{x}.}

\item{output}{a string specifying the type of output requested:
either "percentage" or "proportion". Defaults to proportion.}

\item{output_format}{a string specifying the formatting of the output,
which may include the numerator, the denominator, and percentage/proportion:
one of \code{"p"} (proportion only), \code{"xn"} (numerator and denominator),
\code{"xp"} (numerator and proportion/percentage), or \code{"xnp"}
(numerator/denominator and proportion/percentage)}

\item{digits_proportion}{the digits of precision for proportions
(see \code{table1::round_pad}).}

\item{digits_percent}{the digits of precision for percentages
(see \link[table1:signif_pad]{table1::round_pad}).}

\item{ci_method}{a string containing the type of interval estimate to
compute, if any. See ?\link[binom:binom.confint]{binom::binom.confint}.}

\item{level}{the confidence/credible interval coefficient}

\item{...}{other arguments passed to \link[binom:binom.confint]{binom::binom.confint}}
}
\value{
When a CI is computed, the result is a list, containing the output
and individual components: otherwise, the result is a string.
}
\description{
This is for formatting proportions, percentages and confidence intervals for
reporting. Confidence intervals are calculated using\link[binom:binom.confint]{binom::binom.confint}.
Formatting is performed using \link[table1:signif_pad]{table1::round_pad}.
}
\examples{
# Calculating percentages
in_text_proportion(x = 2, n = 10, output_format = "p")
in_text_proportion(x = 0, n = 10, output_format = "xp")
in_text_proportion(x = 10, n = 10, output_format = "xnp")

in_text_proportion(x = 2, n = 10, output = "proportion", output_format = "p")
in_text_proportion(x = 0, n = 10, output = "proportion", output_format = "xp")
in_text_proportion(x = 10, n = 10, output = "proportion", output_format = "xnp")

in_text_proportion(x = 1, n = 1e4, output_format = "xp", digits_percent = 2)
in_text_proportion(x = 1, n = 1e4, output_format = "xp", digits_percent = 2)

in_text_proportion(x = 0, n = 10, output_format = "p", ci_method = "bayes")
in_text_proportion(x = 0, n = 10, output_format = "p", ci_method = "exact")
}
