% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_whole_number.R
\name{is_whole_number}
\alias{is_whole_number}
\title{Determine whether numeric elements are whole numbers}
\usage{
is_whole_number(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A \code{numeric} \code{vector} or \code{matrix}}

\item{tol}{The precision threshold used to make the determination of a whole
number.}
}
\value{
A \code{logical} vector or matrix, depending on the class of
\code{x}.
}
\description{
Note: this is shamelessly adapted from ?is.integer
}
\examples{
is_whole_number(0)
is_whole_number(0.1)
is_whole_number(0.1^7)
is_whole_number(0.1^24)
is_whole_number(c(0.1^7, 0.1^24))
}
