% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfbr-package.R
\docType{package}
\name{jfbr-package}
\alias{jfbr}
\alias{jfbr-package}
\title{jfbr: What the Package Does (One Line, Title Case)}
\description{
Miscellaneous functions for a statistical workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jbetz-jhu/jfbr}
}

}
\author{
\strong{Maintainer}: Josh Betz \email{jbetz@jhu.edu} (\href{https://orcid.org/0000-0003-4488-9799}{ORCID})

}
\keyword{internal}
