% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_apply_rounding_jfbr.R
\name{stats_apply_rounding_jfbr}
\alias{stats_apply_rounding_jfbr}
\title{Apply rounding to basic descriptive statistics.}
\usage{
stats_apply_rounding_jfbr(
  x,
  digits = 3,
  digits.pct = 1,
  round.median.min.max = TRUE,
  round.integers = TRUE,
  round5up = TRUE,
  rounding.fn = table1::signif_pad,
  ...
)
}
\arguments{
\item{x}{A list, such as that returned by \code{stats.default}.}

\item{digits}{An integer specifying the number of significant digits to keep.}

\item{digits.pct}{An integer specifying the number of digits after the
decimal place for percentages.}

\item{round.median.min.max}{Should rounding applied to median, min and max?}

\item{round.integers}{should rounding be limited to digits to the right of
the decimal point?}

\item{round5up}{Should numbers with 5 as the last digit always be rounded up?
The standard R approach is "go to the even digit" (IEC 60559 standard,
see \code{round}), while some other software (e.g. SAS, Excel) always round
up.}

\item{rounding.fn}{The function to use to do the rounding. Defaults to
\code{signif_pad}.}

\item{...}{Further arguments.}
}
\value{
A list with the same number of elements as \code{x}. The rounded
values will be character (not numeric) and will have 0 padding to ensure
consistent number of significant digits.
}
\description{
This is a slight modification of \link[table1:stats.apply.rounding]{table1::stats.apply.rounding}. For more
information
}
\examples{
library(table1)
stats.default(x = jfbr_test$continuous)
stats.apply.rounding(stats.default(x = jfbr_test$continuous), digits = 3)
stats_jfbr(x = jfbr_test$continuous)
stats_apply_rounding_jfbr(stats_jfbr(x = jfbr_test$continuous), digits = 3)
stats_apply_rounding_jfbr(
  x =
    stats_jfbr(
      x = jfbr_test$continuous,
      custom_quantiles = c(0.47, 0.52)
    ),
  digits = 3
)
}
