% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_jfbr.R
\name{stats_jfbr}
\alias{stats_jfbr}
\title{Compute some basic descriptive statistics.}
\usage{
stats_jfbr(x, custom_quantiles = NULL, quantile_type = 7, ...)
}
\arguments{
\item{x}{A vector of \code{numeric}, \code{factor}, \code{character}, or
\code{logical} values.}

\item{custom_quantiles}{A vector of \code{numeric} values in (0, 1) of
quantiles to compute}

\item{quantile_type}{An integer from 1 to 9, passed as the type argument to
function \code{\link[stats:quantile]{stats::quantile()}}}

\item{...}{Further arguments (ignored).}
}
\value{
A list of statistics. See \code{?table1::stats.default}.
}
\description{
A slight modification of \code{\link[table1:stats.default]{table1::stats.default()}} to allow for custom
quantiles and produce the percentage of non-missing values. For more
information, see \code{?table1::stats.default}.
}
\examples{
stats_jfbr(x = jfbr_test$continuous)
stats_jfbr(x = jfbr_test$continuous, custom_quantiles = c(0.42, 0.57))
}
