% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_numeric.R
\name{table1_numeric}
\alias{table1_numeric}
\title{Summary Functions for table1::table1()}
\usage{
table1_numeric(
  x,
  mean_sd = TRUE,
  median_iqr = TRUE,
  range = TRUE,
  quantiles = NULL,
  na.rm = FALSE,
  quantile_type = 7,
  digits = 2
)
}
\arguments{
\item{x}{A \code{vector} of values passed from \code{\link[table1:table1]{table1::table1()}}}

\item{mean_sd}{A \code{logical} scalar: Compute Mean & SD?}

\item{median_iqr}{A \code{logical} scalar: Compute Median & IQR?}

\item{range}{A \code{logical} scalar: Compute Range?}

\item{quantiles}{A \code{vector} of \code{numeric} values: Quantiles to
compute}

\item{na.rm}{\code{logical} scalar: Remove \code{NA} values?}

\item{quantile_type}{A \code{numeric} scalar, indicating the method of
computing quantiles with \code{\link[stats:quantile]{stats::quantile()}}.}

\item{digits}{A \code{numeric} scalar, indicating the number of significant
figures to include in results.}
}
\value{
A \code{vector} of character-formatted results
}
\description{
The default functions in table1::table1() compute the Mean (SD) and Median
(Min, Max) for continuous values. These summary functions are meant to offer
more flexibility in terms of summaries.
}
\examples{
table1_numeric(
  x = c(1:100, NA),
  quantiles = c(0.05, 0.95)
)
}
