% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_pvalue.R
\name{table1_pvalue}
\alias{table1_pvalue}
\title{Compute hypothesis tests for table1::table1()}
\usage{
table1_pvalue(
  x,
  variable,
  digits = 3,
  test_numeric_2_levels = stats::t.test,
  test_numeric_more_than_2_levels = function(data, formula) {
     data.frame(p.value =
    stats::anova(stats::lm(formula = formula, data = data))$`Pr(>F)`[1])
 },
  test_categorical_2_levels = stats::chisq.test,
  test_categorical_more_than_2_levels = stats::chisq.test,
  ...
)
}
\arguments{
\item{x}{Passed by \code{table1::table1()} - leave empty.}

\item{variable}{Passed by \code{table1::table1()} - leave empty.}

\item{digits}{\code{numeric} scalar: number of digits to use when formatting
p-values}

\item{test_numeric_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has only two levels, and
\code{variable} is numeric. The default is \code{t.test}, alternatives include:
\code{wilcox.test}.}

\item{test_numeric_more_than_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has more than two levels,
and \code{variable} is numeric. The default is the omnibus F-test from
anova(lm(variable ~ group)), alternatives include: \code{kruskal.test}.}

\item{test_categorical_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has only two levels and
\code{variable} is categorical. The default is \code{chisq.test}, alternatives
include: \code{fisher.test}.}

\item{test_categorical_more_than_2_levels}{a \code{function} that computes a
hypothesis test using a formula (variable ~ group) where group has more than
two levels and \code{variable} is categorical. The default is \code{chisq.test},
alternatives include: \code{fisher.test}.}

\item{...}{Other arguments: not currently used.}
}
\value{
A \code{numeric} scalar containing the p-value from the specified
hypothesis tests.
}
\description{
table1::table1() has the ability to add a column for hypothesis test
p-values: see ?table1::table1.
}
\examples{
# To be added
}
