% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_render.R
\name{table1_render}
\alias{table1_render}
\title{Wrapper for render in \link[table1:table1]{table1::table1}}
\usage{
table1_render(numeric = table1_numeric, categorical = table1_categorical, ...)
}
\arguments{
\item{numeric}{A function for computing summaries for numeric variables
(e.g. \link[table1:render.categorical.default]{table1::render.categorical.default})}

\item{categorical}{A function for computing summaries for categorical
variables (e.g. \link[table1:render.continuous.default]{table1::render.continuous.default})}

\item{...}{Other arguments passed onto \link[table1:render.default]{table1::render.default}}
}
\value{
The evaluated results
}
\description{
Wrapper for render in \link[table1:table1]{table1::table1}
}
\examples{
table1::table1(
  x = ~ numbers + continuous + binary + ordered +
    binary_factor + categorical | three_level_group,
  data = jfbr_test,
  render = table1_render
)

table1::table1(
  x = ~ numbers + continuous + binary + ordered +
    binary_factor + categorical | three_level_group,
  data = jfbr_test,
  render =
      numeric = render.continuous.default,
      categorical = render.categorical.default,
      ...
    )
)
}
