% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_numeric.R
\name{table1_numeric}
\alias{table1_numeric}
\title{Summary Functions for table1::table1() for Numeric Data}
\usage{
table1_numeric(
  x,
  mean_sd = TRUE,
  median_iqr = TRUE,
  range = TRUE,
  quantiles = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{vector} of values passed from \code{\link[table1:table1]{table1::table1()}}}

\item{mean_sd}{A \code{logical} scalar: Compute Mean & SD?}

\item{median_iqr}{A \code{logical} scalar: Compute Median & IQR?}

\item{range}{A \code{logical} scalar: Compute Range?}

\item{quantiles}{A \code{vector} of \code{numeric} values: Quantiles to
compute}

\item{...}{Arguments passed to \code{\link[=stats_jfbr]{stats_jfbr()}}}
}
\value{
A \code{vector} of character-formatted results
}
\description{
The default functions in table1::table1() compute the Mean (SD) and Median
(Min, Max) for continuous values. These summary functions are meant to offer
more flexibility in terms of summaries.
}
\examples{

table1_numeric(
  x = c(1:100, NA),
  quantiles = c(0.05, 0.95)
)

data(jfbr_test)
library(table1)

  table1(
    x = ~ continuous + numbers | two_level_group,
    data = jfbr_test,
    render.continuous = table1_numeric
  )
}
