% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_pvalue.R
\name{table1_pvalue}
\alias{table1_pvalue}
\title{Compute hypothesis tests for table1::table1()}
\usage{
table1_pvalue(
  x,
  variable,
  test_numeric_2_levels = stats::t.test,
  test_numeric_more_than_2_levels = function(data, formula) {
     data.frame(p.value =
    stats::anova(stats::lm(formula = formula, data = data))$`Pr(>F)`[1])
 },
  test_categorical_2_levels = stats::chisq.test,
  test_categorical_more_than_2_levels = stats::chisq.test,
  digits = 3
)
}
\arguments{
\item{x}{A named \code{list} of values - Typically passed from
table1::table1().}

\item{variable}{a \code{character} scalar indicating the grouping variable
name - Typically passed from table1::table1().}

\item{test_numeric_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has only two levels, and
\code{variable} is numeric. The default is \code{t.test}, alternatives include:
\code{wilcox.test}.}

\item{test_numeric_more_than_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has more than two levels,
and \code{variable} is numeric. The default is the omnibus F-test from
anova(lm(variable ~ group)), alternatives include: \code{kruskal.test}.}

\item{test_categorical_2_levels}{a \code{function} that computes a hypothesis
test using a formula (variable ~ group) where group has only two levels and
\code{variable} is categorical. The default is \code{chisq.test}, alternatives
include: \code{fisher.test}.}

\item{test_categorical_more_than_2_levels}{a \code{function} that computes a
hypothesis test using a formula (variable ~ group) where group has more than
two levels and \code{variable} is categorical. The default is \code{chisq.test},
alternatives include: \code{fisher.test}.}

\item{digits}{\code{numeric} scalar: number of digits to use when formatting
p-values}
}
\value{
A \code{numeric} scalar containing the p-value from the specified
hypothesis tests.
}
\description{
table1::table1() has the ability to add a column for hypothesis test
p-values: see ?table1::table1.
}
\examples{

set.seed(12345)
n_obs <- 200

jfbr_test <-
  data.frame(
    numbers = 1:n_obs,
    continuous = runif(n = n_obs),
    binary = rbinom(n = n_obs, size = 1, prob = 0.5),
    binary_factor =
      factor(
        x = rbinom(n = n_obs, size = 1, prob = 0.5),
        levels = 0:1,
        labels = c("0. No", "1. Yes")
      ),
    categorical = factor(sample(x = 1:4, size = n_obs, replace = TRUE)),
    ordered = ordered(sample(x = 1:4, size = n_obs, replace = TRUE))
  )


table1::table1(
  x = ~ numbers + continuous + binary + ordered | categorical,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" = table1_pvalue)
)

table1::table1(
  x = ~ numbers + continuous + binary + ordered | binary_factor,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" =
           function(x, value) table1_pvalue(
             x = x,
             variable = variable,
             test_numeric_2_levels = wilcox.test,
             test_numeric_more_than_2_levels = kruskal.test,
             test_categorical_2_levels = fisher.test,
             test_categorical_more_than_2_levels = fisher.test
           )
    )
)
}
